server/src/services/flask/requirements.txt
flask
flask_cors
python-dotenv
openai
pyttsx3
supabase

